Rename SearchBar, SearchResults components specific to skills
